+++ tar/storage/storage_internal.h
@@ -23,6 +23,6 @@ int storage_transaction_start_delete(NETPACKET_CONNECTION *, uint64_t,
  * transaction in ${seqnum}.  Use the key specified by whichkey.
  */
 int storage_transaction_start_fsck(NETPACKET_CONNECTION *, uint64_t,
-    uint8_t[32], int);
+    uint8_t[32], uint8_t);
 
 #endif /* !_STORAGE_INTERNAL_H_ */
